Add sm3-yescrypt backend and tests.
sm3-yescrypt uses the output from the yescrypt hashing method in
place of a hmac message.  Thus, the yescrypt crypto properties
are superseeded by the ShangMi 3 hash function with a 256 bit
digest.

This hashing method is useful in applications that need modern
password hashing methods, but require to rely on cryptographic
properties that have been approved by the Chinese Office of State
Commercial Cryptography Administration (OSCCA).

ShangMi 3 is a cryptographic hash function used in the Chinese
National Standard.  It was published by the National Cryptography
Administration (Chinese: 国家密码管理局) on 2010-12-17 as
"GM/T 0004-2012: SM3 cryptographic hash algorithm".

SM3 is used for implementing digital signatures, message
authentication codes, and pseudorandom number generators.  The
algorithm is public and is considered similar to SHA-256 in
security and efficiency.
SM3 is also used with Transport Layer Security.

SM3 is defined in each of:

  * GM/T 0004-2012:   SM3 cryptographic hash algorithm
  * GB/T 32905-2016:  Information security techniques
                      SM3 cryptographic hash algorithm
  * ISO/IEC 10118-3:2018:
                      IT Security techniques—Hash-functions
                      Part 3: Dedicated hash-functions
  * IETF RFC draft-sca-cfrg-sm3-02

See:  https://datatracker.ietf.org/doc/html/draft-sca-cfrg-sm3