lib/crypt.c: Do not overwrite the output buffer too early.
Allow passing a pointer returned by crypt(3) previously as the input
for a subsequent call to crypt(3), adapt the manpage accordingly,
and implement a simple testcase for such a scenario.  This also
applies to the other variants of the crypt function, when the same
data object is used for subsequent calls.

Alter the resize process of the data object by crypt_ra(3), to keep
the literals passed with the phrase and the setting parameter during
the runtime of the function, if those are stored within the initially
passed data object.

Also apply the same semantics to the crypt_gensalt(3) function family,
and the internally used make_failure_token() helper.

See: https://www.openwall.com/lists/yescrypt/2025/04/08/1

Fixes #209.