test/getrandom-fallbacks: fix open shim types
In C, functions taking a variable number of arguments are not guaranteed to have
the same calling convention as functions taking a fixed number of arguments
[^0].  Since `open()` and `open64()` are defined as taking a variable number of
arguments, the `__wrap_open()` and `__wrap_open64()` mocks must be as well.

[^0]: C23 §6.7.7.4¶14 says: "For two function types to be compatible, both shall
specify compatible return types. Moreover, the parameter type lists shall agree
in the number of parameters and in use of the final ellipsis; corresponding
parameters shall have compatible types."

This fixes the test on my powerpc64el machine building nixpkgs a8d610af3f1.